Podzielono widoki na komponenty, podzielono style css
