+++ package.json
@@ -13,6 +13,7 @@
     "react-dom": "^17.0.2",
     "react-final-form": "^6.5.3",
     "react-icons": "^4.2.0",
+    "react-image-magnifiers": "^1.4.0",
     "react-redux": "^7.2.4",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
