+++ src/Pages/Product/ProductsAside/ProductsCategoriesListItem.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const ProductsCategoriesListItem = ({ name }) => {
+  return (
+    <li className='categories-list__item'>
+      <span>{name}</span>
+    </li>
+  );
+};
+
+export default ProductsCategoriesListItem;
