+++ src/Pages/Product/ProductsAside/ProductsCategoriesListItems.js
@@ -1,0 +1,25 @@
+import React from 'react';
+import ProductsCategoriesListItem from './ProductsCategoriesListItem';
+
+const ProductsCategoriesListItems = () => {
+  return (
+    <ul className='categories-list__items'>
+      {[
+        'Promocje',
+        'Nowości',
+        'Dla niego',
+        'Dla niej',
+        'Dla dziecka',
+        'Wyposażenie turystyczne',
+        'Narty',
+        'Kemping',
+        'Fitness',
+        'Wyprzedaż',
+      ].map((category, idx) => (
+        <ProductsCategoriesListItem key={idx} name={category} />
+      ))}
+    </ul>
+  );
+};
+
+export default ProductsCategoriesListItems;
