+++ src/Style/scss/_account-page.scss
@@ -1,0 +1,5 @@
+.account-page {
+  display: flex;
+  margin-top: 20px;
+  height: calc(100vh - 110px);
+}
