+++ src/Style/scss/_cart-page.scss
@@ -1,0 +1,5 @@
+.cart-page {
+  display: flex;
+  margin-top: 20px;
+  height: calc(100vh - 110px);
+}
