+++ src/Style/scss/_footer.scss
@@ -1,0 +1,191 @@
+.footer {
+  padding-top: 100px;
+  padding-left: 30px;
+  padding-right: 30px;
+  min-height: 250px;
+  background: linear-gradient(rgba(0, 0, 0, 0.8), rgba(0, 0, 0, 0.8)),
+    url('https://img.wallpapersafari.com/desktop/1920/1080/44/22/REgNxl.jpg');
+  background-size: cover;
+  width: 100%;
+  display: flex;
+  align-items: stretch;
+  flex-wrap: wrap;
+
+  &__contact,
+  &__shortcut,
+  &__socials {
+    flex-basis: 33.3333%;
+  }
+
+  &__explanation {
+    display: flex;
+    height: 50px;
+    align-items: center;
+    justify-content: center;
+    flex-basis: 100%;
+    background-color: rgba(0, 0, 0, 0.3);
+    color: darken($white-color, 30%);
+  }
+}
+
+.contact {
+  display: flex;
+  flex-direction: column;
+  flex-wrap: nowrap;
+  padding-left: 10px;
+  padding-right: 10px;
+
+  &__title {
+    color: $white-color;
+    font-size: 1.6rem;
+    margin-top: 0;
+    margin-bottom: 15px;
+  }
+
+  &__line {
+    margin-right: 30px;
+    margin-bottom: 30px;
+    height: 1px;
+    width: 100%;
+    border: 0;
+    background-color: $white-color;
+  }
+
+  &__form {
+    display: flex;
+    justify-content: flex-start;
+    flex-wrap: wrap;
+    margin-bottom: 20px;
+  }
+
+  &__text-input,
+  &__text-area {
+    display: flex;
+    flex-basis: 90%;
+    margin: 10px 0;
+    padding: 10px 5px;
+    background-color: rgba(0, 0, 0, 0.3);
+    border: 3px solid $white-color;
+    color: $white-color;
+  }
+
+  &__text-input::placeholder,
+  &__text-area::placeholder {
+    color: $white-color;
+    opacity: 1;
+  }
+
+  &__text-area:-ms-input-placeholder,
+  &__text-input:-ms-input-placeholder {
+    color: $white-color;
+  }
+
+  &__text-area::-ms-input-placeholder,
+  &__text-input::-ms-input-placeholder {
+    color: $white-color;
+  }
+
+  &__clear-form {
+    background-color: $white-color;
+    color: $black-color;
+    padding: 5px 10px;
+    border: 2px solid $white-color;
+  }
+
+  &__send-form {
+    background-color: rgba(0, 0, 0, 0.2);
+    color: $white-color;
+    margin-right: 10px;
+    padding: 10px 15px;
+    border: 2px solid $white-color;
+  }
+}
+
+.shortcuts {
+  display: flex;
+  flex-direction: column;
+  flex-wrap: nowrap;
+  padding-left: 10px;
+  padding-right: 10px;
+
+  &__title {
+    color: $white-color;
+    font-size: 1.6rem;
+    margin-top: 0;
+    margin-bottom: 15px;
+  }
+
+  &__line {
+    margin-right: 30px;
+    height: 1px;
+    border: 0;
+    width: 100%;
+    background-color: $white-color;
+  }
+}
+
+.social-media {
+  display: flex;
+  flex-direction: column;
+  flex-wrap: nowrap;
+  padding-left: 10px;
+  padding-right: 10px;
+  height: 100%;
+
+  &__title {
+    color: $white-color;
+    font-size: 1.6rem;
+    margin-top: 0;
+    margin-bottom: 15px;
+  }
+
+  &__line {
+    margin-right: 30px;
+    height: 1px;
+    border: 0;
+    width: 100%;
+    background-color: $white-color;
+  }
+
+  &__container {
+    display: flex;
+    height: 60%;
+    justify-content: center;
+    align-items: center;
+
+    & a {
+      color: $white-color;
+      font-size: 3.5rem;
+      margin-left: 20px;
+      margin-right: 20px;
+    }
+  }
+}
+
+.text-input,
+.text-area {
+  display: flex;
+  flex-wrap: wrap;
+  flex-basis: 100%;
+
+  &__label {
+    flex-basis: 100%;
+    font-size: 0.75rem;
+    color: $secondary-color;
+    margin-bottom: 10px;
+  }
+}
+
+.shortcuts {
+  &__list {
+    margin-top: 40px;
+    padding-left: 10px;
+    list-style-type: none;
+    color: white;
+
+    & li {
+      margin-top: 15px;
+      margin-bottom: 15px;
+    }
+  }
+}
