+++ src/Style/scss/_products-page.scss
@@ -1,0 +1,290 @@
+.products-page {
+  display: flex;
+  flex-direction: row;
+  flex-wrap: nowrap;
+  margin: 100px 40px 0 40px;
+  min-height: calc(100vh - 110px);
+
+  & .categories-list {
+    flex-basis: 30%;
+  }
+
+  & .products-view {
+    flex-basis: 70%;
+  }
+}
+
+.categories-list {
+  padding-right: 80px;
+  &__header {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    font-size: 1.8rem;
+    font-weight: 300;
+  }
+
+  &__header::after {
+    content: '';
+    display: flex;
+    background-color: $secondary-color;
+    flex-basis: 35%;
+    height: 2px;
+  }
+
+  &__items {
+    list-style-type: none;
+    padding-right: 30px;
+    padding-left: 20px;
+  }
+
+  &__item {
+    span {
+      display: block;
+      padding-left: 10px;
+      padding-top: 15px;
+      padding-bottom: 15px;
+    }
+  }
+
+  &__item:hover {
+    background-color: $primary-color;
+    color: $white-color;
+
+    &::after {
+      background-color: $white-color;
+    }
+  }
+
+  &__item::after {
+    content: '';
+    display: block;
+    position: relative;
+    bottom: 0;
+    height: 2px;
+    width: 100%;
+    background-color: rgba(0, 0, 0, 0.1);
+  }
+
+  &__filters {
+    margin-top: 100px;
+  }
+}
+
+.products-view {
+  display: flex;
+  flex-direction: column;
+  flex-wrap: nowrap;
+
+  &__headline {
+    display: flex;
+    align-items: center;
+    flex-direction: row;
+    flex-wrap: nowrap;
+  }
+
+  &__view-btn {
+    background: none;
+    border: none;
+    font-size: 2rem;
+
+    &--active {
+      color: $secondary-color;
+    }
+  }
+
+  &__header {
+    display: flex;
+    align-items: center;
+    flex-basis: 70%;
+    font-size: 1.8rem;
+    font-weight: 300;
+  }
+
+  &__header::after {
+    content: '';
+    display: flex;
+    background-color: $secondary-color;
+    min-width: 20%;
+    max-width: 100%;
+    margin-left: 20px;
+    height: 2px;
+  }
+
+  &__controls {
+    display: flex;
+    justify-content: space-evenly;
+    flex-basis: 30%;
+  }
+
+  &__products {
+    display: grid;
+    grid-template-rows: auto;
+    margin-bottom: 50px;
+
+    &--grid {
+      grid-template-columns: repeat(3, 1fr);
+      gap: 15px 15px;
+    }
+
+    &--list {
+      grid-template-columns: repeat(1, 1fr);
+      row-gap: 15px;
+    }
+  }
+
+  &__product {
+    border: 1px solid rgba(0, 0, 0, 0.1);
+  }
+}
+
+.product {
+  $this: &;
+  display: flex;
+
+  &--grid {
+    padding-left: 5px;
+    padding-right: 5px;
+    flex-direction: column;
+
+    #{$this}__image {
+      padding: 30px;
+    }
+
+    #{$this}__info {
+      padding-left: 10px;
+      padding-right: 10px;
+    }
+
+    #{$this}__name {
+      margin-top: 0;
+      margin-bottom: 0;
+      font-size: 1.8rem;
+      font-weight: 500;
+    }
+
+    #{$this}__price {
+      font-size: 1.3rem;
+      color: $secondary-color;
+      margin-top: 0;
+      margin-bottom: 10px;
+      font-weight: 400;
+    }
+
+    #{$this}__desc {
+      margin-top: 0;
+      font-weight: 100;
+    }
+
+    #{$this}__add-to-cart {
+      border: 0px;
+      padding-top: 10px;
+      padding-bottom: 10px;
+      margin-bottom: 5px;
+      background-color: rgba(0, 0, 0, 0.05);
+      cursor: pointer;
+    }
+  }
+
+  &--list {
+    flex-direction: row;
+
+    #{$this}__image {
+      max-width: 200px;
+      padding: 20px;
+    }
+
+    #{$this}__name {
+      margin-top: 20px;
+      margin-bottom: 0;
+      font-size: 1.8rem;
+      font-weight: 500;
+    }
+
+    #{$this}__price {
+      font-size: 1.3rem;
+      color: $secondary-color;
+      margin-top: 0;
+      margin-bottom: 10px;
+      font-weight: 400;
+    }
+
+    #{$this}__desc {
+      margin-top: 0;
+      font-weight: 100;
+    }
+
+    #{$this}__add-to-cart {
+      border: 0px;
+      padding-left: 10px;
+      padding-right: 10px;
+      background-color: rgba(0, 0, 0, 0.05);
+      cursor: pointer;
+    }
+  }
+}
+
+.filters {
+  display: grid;
+  grid-template-columns: 1;
+  grid-template-rows: auto;
+  row-gap: 10px;
+  padding-left: 10px;
+  padding-right: 10px;
+
+  &__price {
+    border: 1px solid rgba(0, 0, 0, 0.1);
+    display: flex;
+    flex-direction: row;
+    justify-content: space-evenly;
+    flex-wrap: wrap;
+    padding-top: 20px;
+    padding-bottom: 20px;
+
+    & input {
+      width: 100px;
+      border: none;
+      background-color: rgba(0, 0, 0, 0.05);
+      margin: 5px;
+      padding: 10px;
+    }
+  }
+
+  &__producer,
+  &__gender {
+    border: 1px solid rgba(0, 0, 0, 0.1);
+    display: flex;
+    flex-direction: row;
+    justify-content: space-evenly;
+    align-items: center;
+    flex-wrap: nowrap;
+    padding-top: 20px;
+    padding-bottom: 20px;
+
+    & h4 {
+      flex-basis: 30%;
+      margin-left: 10px;
+      margin-right: 10px;
+    }
+
+    & form {
+      display: flex;
+      flex-direction: column;
+    }
+  }
+
+  &__apply-btn {
+    background-color: $primary-color;
+    color: $white-color;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    padding-top: 20px;
+    padding-bottom: 20px;
+    font-size: 1.1rem;
+  }
+
+  &__filter-icon {
+    margin-left: 10px;
+  }
+}
